requirements.txt
Pillow>=10.0.0
imagecodecs>=2023.1.0
